+++ editor/editor-plugin-avatar-group/src/ui/colored-avatar-item.tsx
@@ -6,7 +6,8 @@
 import { jsx } from '@emotion/react';
 
 import type { ExtractInjectionAPI } from '@atlaskit/editor-common/types';
-import { R100 } from '@atlaskit/theme/colors';
+import { N900, R100 } from '@atlaskit/theme/colors';
+import { token } from '@atlaskit/tokens';
 
 import type { AvatarGroupPlugin } from '../avatarGroupPluginType';
 
@@ -14,16 +15,27 @@ import { badge } from './styles';
 
 export interface ColoredAvatarItemProps {
 	sessionId: string;
+	presenceId?: string;
 	name: string;
 	api: ExtractInjectionAPI<AvatarGroupPlugin> | undefined;
 }
 
 export const ColoredAvatarItem = (props: ColoredAvatarItemProps) => {
-	const color = props.api?.collabEdit?.actions?.getAvatarColor(props.sessionId).color;
+	const participantColor = props.api?.collabEdit?.actions?.getAvatarColor(
+		props.presenceId ?? props.sessionId,
+	) ?? {
+		backgroundColor: token('color.background.accent.red.subtle', R100),
+		textColor: token('color.text.accent.gray.bolder', N900),
+	};
+
 	const avatar = props.name.substr(0, 1).toUpperCase();
+
 	return (
-		// eslint-disable-next-line @atlaskit/design-system/consistent-css-prop-usage, @atlaskit/ui-styling-standard/no-imported-style-values -- Ignored via go/DSP-18766
-		<div css={badge(color || R100)} data-testid="editor-collab-badge">
+		<div
+			// eslint-disable-next-line @atlaskit/design-system/consistent-css-prop-usage, @atlaskit/ui-styling-standard/no-imported-style-values -- Ignored via go/DSP-18766
+			css={badge(participantColor?.backgroundColor, participantColor?.textColor)}
+			data-testid="editor-collab-badge"
+		>
 			{avatar}
 		</div>
 	);