+++ editor/editor-common/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @atlaskit/editor-common
 
+## 97.3.0
+
+### Minor Changes
+
+- [#98820](https://stash.atlassian.com/projects/CONFCLOUD/repos/confluence-frontend/pull-requests/98820)
+  [`9f7dc3ba5d256`](https://stash.atlassian.com/projects/CONFCLOUD/repos/confluence-frontend/commits/9f7dc3ba5d256) -
+  [ux] Update the potential list of colors for avatar badges and telepointers within the editor.
+  Also add support for using presenceId over sessionId, when presenceId exists, in order to make the
+  avatar badge and telepointer color selection.
+
 ## 97.2.0
 
 ### Minor Changes